12 feature irida 10 add the ability to display documentation for this telegram bot

+++ src/main/java/com/github/iridatelegrambot/service/InvoiceServiceImpl.java
@@ -2,6 +2,8 @@ package com.github.iridatelegrambot.service;
 
 import com.github.iridatelegrambot.entity.Invoice;
 import com.github.iridatelegrambot.repository.InvoiceRepository;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -11,6 +13,7 @@ import java.util.Optional;
 public class InvoiceServiceImpl implements InvoiceService {
 
     private final InvoiceRepository invoiceRepository;
+
     @Autowired
     public InvoiceServiceImpl(InvoiceRepository invoiceRepository) {
         this.invoiceRepository = invoiceRepository;
@@ -32,8 +35,11 @@ public class InvoiceServiceImpl implements InvoiceService {
     }
 
     @Override
-    public void detele(int id) {
-        Invoice invoice = getInvoiceById(id).get();
-        invoiceRepository.delete(invoice);
+    public void delete(int id) {
+        Optional<Invoice> invoiceOptional = getInvoiceById(id);
+        if(invoiceOptional.isPresent()) {
+            Invoice invoice = invoiceOptional.get();
+            invoiceRepository.delete(invoice);
+        }
     }
 }
